orchestrator/requirements.txt
# Essential Flask packages
Flask
gunicorn
python-dotenv

# AutoGen dependencies # <--- 以下の行を追加
autogen-agentchat>=0.7.0
autogen-ext[openai]>=0.7.0
openai>=1.101.0